+++ plugins/org.eclipse.elk.graph.json/tests/all/mod.rs
@@ -2,6 +2,7 @@ mod common;
 mod edge_coords_test;
 mod edges_test;
 mod export_test;
+mod fixtures_layout_test;
 mod graph_test;
 mod id_test;
 mod individual_spacings_test;